+++ bonsai/proto/generator_simulator_api_pb2.py
@@ -20,10 +20,9 @@ DESCRIPTOR = _descriptor.FileDescriptor(
   name='bonsai/proto/generator_simulator_api.proto',
   package='bonsai.proto.generator_simulator_api',
   syntax='proto3',
-  serialized_pb=_b('\n*bonsai/proto/generator_simulator_api.proto\x12$bonsai.proto.generator_simulator_api\x1a google/protobuf/descriptor.proto\"]\n\x14SimulationSourceData\x12\r\n\x05state\x18\x01 \x01(\x0c\x12\x0e\n\x06reward\x18\x02 \x01(\x02\x12\x10\n\x08terminal\x18\x03 \x01(\x08\x12\x14\n\x0c\x61\x63tion_taken\x18\x04 \x01(\x0c\"&\n\x0cRegisterData\x12\x16\n\x0esimulator_name\x18\x01 \x01(\t\"\xd9\x02\n\x11SimulatorToServer\x12Y\n\x0cmessage_type\x18\x01 \x01(\x0e\x32\x43.bonsai.proto.generator_simulator_api.SimulatorToServer.MessageType\x12I\n\rregister_data\x18\x02 \x01(\x0b\x32\x32.bonsai.proto.generator_simulator_api.RegisterData\x12N\n\nstate_data\x18\x03 \x03(\x0b\x32:.bonsai.proto.generator_simulator_api.SimulationSourceData\x12\x0e\n\x06sim_id\x18\x04 \x01(\x07\">\n\x0bMessageType\x12\x0b\n\x07UNKNOWN\x10\x00\x12\x0c\n\x08REGISTER\x10\x01\x12\t\n\x05READY\x10\x02\x12\t\n\x05STATE\x10\x03\"\xdc\x01\n\x17\x41\x63knowledgeRegisterData\x12;\n\x11properties_schema\x18\x01 \x01(\x0b\x32 .google.protobuf.DescriptorProto\x12\x37\n\routput_schema\x18\x02 \x01(\x0b\x32 .google.protobuf.DescriptorProto\x12;\n\x11prediction_schema\x18\x03 \x01(\x0b\x32 .google.protobuf.DescriptorProto\x12\x0e\n\x06sim_id\x18\x04 \x01(\x07\"\x81\x01\n\x11SetPropertiesData\x12\x1a\n\x12\x64ynamic_properties\x18\x01 \x01(\x0c\x12\x13\n\x0breward_name\x18\x02 \x01(\t\x12;\n\x11prediction_schema\x18\x03 \x01(\x0b\x32 .google.protobuf.DescriptorProto\",\n\x0ePredictionData\x12\x1a\n\x12\x64ynamic_prediction\x18\x01 \x01(\x0c\"\xfe\x03\n\x11ServerToSimulator\x12Y\n\x0cmessage_type\x18\x01 \x01(\x0e\x32\x43.bonsai.proto.generator_simulator_api.ServerToSimulator.MessageType\x12`\n\x19\x61\x63knowledge_register_data\x18\x02 \x01(\x0b\x32=.bonsai.proto.generator_simulator_api.AcknowledgeRegisterData\x12T\n\x13set_properties_data\x18\x03 \x01(\x0b\x32\x37.bonsai.proto.generator_simulator_api.SetPropertiesData\x12M\n\x0fprediction_data\x18\x04 \x03(\x0b\x32\x34.bonsai.proto.generator_simulator_api.PredictionData\"\x86\x01\n\x0bMessageType\x12\x0b\n\x07UNKNOWN\x10\x00\x12\x18\n\x14\x41\x43KNOWLEDGE_REGISTER\x10\x01\x12\x12\n\x0eSET_PROPERTIES\x10\x02\x12\t\n\x05START\x10\x03\x12\x08\n\x04STOP\x10\x04\x12\x0e\n\nPREDICTION\x10\x05\x12\t\n\x05RESET\x10\x06\x12\x0c\n\x08\x46INISHED\x10\x07\x62\x06proto3')
+  serialized_pb=_b('\n*bonsai/proto/generator_simulator_api.proto\x12$bonsai.proto.generator_simulator_api\x1a google/protobuf/descriptor.proto\"]\n\x14SimulationSourceData\x12\r\n\x05state\x18\x01 \x01(\x0c\x12\x0e\n\x06reward\x18\x02 \x01(\x02\x12\x10\n\x08terminal\x18\x03 \x01(\x08\x12\x14\n\x0c\x61\x63tion_taken\x18\x04 \x01(\x0c\"&\n\x0cRegisterData\x12\x16\n\x0esimulator_name\x18\x01 \x01(\t\"\xd9\x02\n\x11SimulatorToServer\x12Y\n\x0cmessage_type\x18\x01 \x01(\x0e\x32\x43.bonsai.proto.generator_simulator_api.SimulatorToServer.MessageType\x12I\n\rregister_data\x18\x02 \x01(\x0b\x32\x32.bonsai.proto.generator_simulator_api.RegisterData\x12N\n\nstate_data\x18\x03 \x03(\x0b\x32:.bonsai.proto.generator_simulator_api.SimulationSourceData\x12\x0e\n\x06sim_id\x18\x04 \x01(\x07\">\n\x0bMessageType\x12\x0b\n\x07UNKNOWN\x10\x00\x12\x0c\n\x08REGISTER\x10\x01\x12\t\n\x05READY\x10\x02\x12\t\n\x05STATE\x10\x03\"\xdc\x01\n\x17\x41\x63knowledgeRegisterData\x12;\n\x11properties_schema\x18\x01 \x01(\x0b\x32 .google.protobuf.DescriptorProto\x12\x37\n\routput_schema\x18\x02 \x01(\x0b\x32 .google.protobuf.DescriptorProto\x12;\n\x11prediction_schema\x18\x03 \x01(\x0b\x32 .google.protobuf.DescriptorProto\x12\x0e\n\x06sim_id\x18\x04 \x01(\x07\"\x81\x01\n\x11SetPropertiesData\x12\x1a\n\x12\x64ynamic_properties\x18\x01 \x01(\x0c\x12\x13\n\x0breward_name\x18\x02 \x01(\t\x12;\n\x11prediction_schema\x18\x03 \x01(\x0b\x32 .google.protobuf.DescriptorProto\",\n\x0ePredictionData\x12\x1a\n\x12\x64ynamic_prediction\x18\x01 \x01(\x0c\"\x8e\x04\n\x11ServerToSimulator\x12Y\n\x0cmessage_type\x18\x01 \x01(\x0e\x32\x43.bonsai.proto.generator_simulator_api.ServerToSimulator.MessageType\x12`\n\x19\x61\x63knowledge_register_data\x18\x02 \x01(\x0b\x32=.bonsai.proto.generator_simulator_api.AcknowledgeRegisterData\x12T\n\x13set_properties_data\x18\x03 \x01(\x0b\x32\x37.bonsai.proto.generator_simulator_api.SetPropertiesData\x12M\n\x0fprediction_data\x18\x04 \x03(\x0b\x32\x34.bonsai.proto.generator_simulator_api.PredictionData\x12\x0e\n\x06sim_id\x18\x05 \x01(\x07\"\x86\x01\n\x0bMessageType\x12\x0b\n\x07UNKNOWN\x10\x00\x12\x18\n\x14\x41\x43KNOWLEDGE_REGISTER\x10\x01\x12\x12\n\x0eSET_PROPERTIES\x10\x02\x12\t\n\x05START\x10\x03\x12\x08\n\x04STOP\x10\x04\x12\x0e\n\nPREDICTION\x10\x05\x12\t\n\x05RESET\x10\x06\x12\x0c\n\x08\x46INISHED\x10\x07\x62\x06proto3')
   ,
   dependencies=[google_dot_protobuf_dot_descriptor__pb2.DESCRIPTOR,])
-_sym_db.RegisterFileDescriptor(DESCRIPTOR)
 
 
 
@@ -98,8 +97,8 @@ _SERVERTOSIMULATOR_MESSAGETYPE = _descriptor.EnumDescriptor(
   ],
   containing_type=None,
   options=None,
-  serialized_start=1379,
-  serialized_end=1513,
+  serialized_start=1395,
+  serialized_end=1529,
 )
 _sym_db.RegisterEnumDescriptor(_SERVERTOSIMULATOR_MESSAGETYPE)
 
@@ -403,6 +402,13 @@ _SERVERTOSIMULATOR = _descriptor.Descriptor(
       message_type=None, enum_type=None, containing_type=None,
       is_extension=False, extension_scope=None,
       options=None),
+    _descriptor.FieldDescriptor(
+      name='sim_id', full_name='bonsai.proto.generator_simulator_api.ServerToSimulator.sim_id', index=4,
+      number=5, type=7, cpp_type=3, label=1,
+      has_default_value=False, default_value=0,
+      message_type=None, enum_type=None, containing_type=None,
+      is_extension=False, extension_scope=None,
+      options=None),
   ],
   extensions=[
   ],
@@ -417,7 +423,7 @@ _SERVERTOSIMULATOR = _descriptor.Descriptor(
   oneofs=[
   ],
   serialized_start=1003,
-  serialized_end=1513,
+  serialized_end=1529,
 )
 
 _SIMULATORTOSERVER.fields_by_name['message_type'].enum_type = _SIMULATORTOSERVER_MESSAGETYPE
@@ -440,6 +446,7 @@ DESCRIPTOR.message_types_by_name['AcknowledgeRegisterData'] = _ACKNOWLEDGEREGIST
 DESCRIPTOR.message_types_by_name['SetPropertiesData'] = _SETPROPERTIESDATA
 DESCRIPTOR.message_types_by_name['PredictionData'] = _PREDICTIONDATA
 DESCRIPTOR.message_types_by_name['ServerToSimulator'] = _SERVERTOSIMULATOR
+_sym_db.RegisterFileDescriptor(DESCRIPTOR)
 
 SimulationSourceData = _reflection.GeneratedProtocolMessageType('SimulationSourceData', (_message.Message,), dict(
   DESCRIPTOR = _SIMULATIONSOURCEDATA,